Update Client.java
I don't think it is necessary to copy the data array even for readability reasons because the actual syntax p.getData() is pretty straight forward anyway. I have commented it out just in case someone who wrote it has actually a valid reason for doing it this way that i dont see atm. I know it is a small difference but I thought I will throw that in so I it seems that I am doing something at least for now. :P